% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.bray_curtis}
\alias{.bray_curtis}
\title{Bray-Curtis distance matrix}
\usage{
.bray_curtis(x)
}
\arguments{
\item{x}{Numeric matrix (rows = items, columns = features).}
}
\value{
A \code{dist} object.
}
\description{
Bray-Curtis distance matrix
}
\keyword{internal}
