% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.find_at_rank}
\alias{.find_at_rank}
\title{Find sequences matching a taxonomy value at a given rank}
\usage{
.find_at_rank(taxa, rank_name, values)
}
\arguments{
\item{taxa}{Character matrix of taxonomy assignments.}

\item{rank_name}{Character. Column name in \code{taxa}.}

\item{values}{Character vector. Values to match.}
}
\value{
Character vector of matching row names.
}
\description{
Find sequences matching a taxonomy value at a given rank
}
\keyword{internal}
